+++ externals/sword/src/utilfuns/url.cpp
@@ -2,7 +2,7 @@
  *
  *  url.cpp -	code for an URL parser utility class
  *
- * $Id: url.cpp 2980 2013-09-14 21:51:47Z scribe $
+ * $Id: url.cpp 3439 2016-10-23 08:32:02Z scribe $
  *
  * Copyright 2004-2013 CrossWire Bible Society (http://www.crosswire.org)
  *	CrossWire Bible Society
@@ -233,8 +233,8 @@ const SWBuf URL::encode(const char *urlText) {
 	url = urlText;
 	
 	SWBuf buf;
-	const int length = url.length();
-	for (int i = 0; i < length; i++) { //fill "buf"
+	const long length = url.length();
+	for (long i = 0; i < length; i++) { //fill "buf"
 		const char& c = url[i];
 		buf.append( ((m[c].length()) ? m[c] : SWBuf(c)) );
 	}
@@ -249,7 +249,7 @@ const SWBuf URL::decode(const char *encoded) {
 	text = encoded;	
 
 	SWBuf decoded;	
-	const int length = text.length();
+	const long length = text.length();
 	int i = 0;
 	
 	while (i < length) {
